Remove prisma test from CI, fix withDataUri
