+++ packages/uploads/package.json
@@ -34,7 +34,8 @@
     "check:package": "concurrently npm:check:attw yarn publint",
     "setup:test": "npx prisma migrate reset -f --schema src/__tests__/unit-test-schema.prisma",
     "test": "vitest run",
-    "test:watch": "vitest watch"
+    "test:watch": "vitest watch",
+    "test:prismaExtension": "PRISMA_EXTENSION_TESTS=true vitest run"
   },
   "dependencies": {
     "@redwoodjs/project-config": "workspace:*",
