+++ packages/uploads/vitest.config.mts
@@ -6,7 +6,6 @@ export default defineConfig({
     deps: {
       interopDefault: false,
     },
-    globalSetup: ['vitest.setup.mts'],
   },
 
 })
